pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["py_news*"]

[project]
name = "py-news-m"
version = "0.1.0"
description = "Local-first news article ingestion and retrieval scaffold"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "py-news-m contributors" }]
dependencies = [
  "beautifulsoup4>=4.12",
  "click>=8.1",
  "fastapi>=0.115",
  "httpx>=0.27",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "uvicorn>=0.30",
]

[project.scripts]
py-news = "py_news.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
